chore: Refactor reporting methods to ReportAnalyzer
